requirements.txt
streamlit==1.36.0
requests==2.31.0
streamlit-agraph==0.0.45
networkx==3.2
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
pyvis==0.3.2
plotly==5.18.0
arxiv==1.4.8
reportlab==4.0.5
